backend/code/requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
matplotlib
numpy
networkx

